# docs/contracts/openapi-plans-v3.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Environment Studio hosted v3 plans",
    "version": "0.3.1",
    "description": "Twenty-four fixed-v3 routes including semantic commands, explicit materialization, document inventory, physical entities, structural decisions, binding pages, document comparison, complete locations and computed groups/contributors. Actual compiler publication remains incomplete; test publication witnesses do not qualify runtime availability. See hosted-plan-http-v3.md, hosted-plan-commands-v3.md, hosted-plan-materialization-v3.md hosted-plan-physical-views-v3.md hosted-plan-structural-views-v3.md hosted-plan-document-views-v3.md and hosted-plan-computed-views-v3.md for original session authority, shared admission, bounded transfers and early no-body controller refusals. The guarded package candidate endpoint provides an unqualified ZIP download only; no production export or readback endpoint is provided."
  },
  "paths": {
...
    "/api/v3/plans/{planId}/profile-captures": {
      "post": {
        "operationId": "capturePlanProfileV3",
        "summary": "Capture verified original physical-only schema3 profile",
        "description": "See hosted-plan-workflow-v3.md. Original V3 owner, full view admission and the shared semantic transfer precede input; original authority and full proofs remain checked through owned output (128 MiB, original 30 seconds). The closed capture body permits 64 MiB under the original 30-second clock and requires a complete physical mapping of the original observation, at most 20000 entries, with explicit value-free slot IDs and labels. Capture returns physical-only schema 3 JSON source, at most 1 MiB UTF8, with its exact definition reference. Capture does not save or publish a profile; use the existing workspace draft/history APIs explicitly.",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "parameters": [
          {
            "name": "planId",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Uuid"
            }
          },
          {
            "name": "X-CSRF-TOKEN",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "../../schemas/plan-workflow-v3.schema.json#/$defs/captureRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Owned complete result or explicit computed-rule page; no export authority",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "../../schemas/plan-workflow-v3.schema.json#/$defs/captureResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error400"
          },
          "401": {
            "$ref": "#/components/responses/Error401"
          },
          "403": {
            "$ref": "#/components/responses/Error403"
          },
          "404": {
            "$ref": "#/components/responses/Error404"
          },
          "409": {
            "$ref": "#/components/responses/Error409"
          },
          "413": {
            "$ref": "#/components/responses/Error413"
          },
          "422": {
            "$ref": "#/components/responses/Error422"
          },
          "429": {
            "$ref": "#/components/responses/Error429"
          },
          "503": {
            "$ref": "#/components/responses/Error503"
          },
          "500": {
            "$ref": "#/components/responses/Error500"
          }
        }
      }
    },
    "/api/v3/plans/{planId}/profile-previews": {
      "post": {
        "operationId": "previewPlanProfileV3",
        "summary": "Inspect full qualified v3 profile closure and affected derivations",
        "description": "See hosted-plan-workflow-v3.md. Original V3 owner, full view admission and the shared semantic transfer precede input; original authority and full proofs remain checked through owned output (128 MiB, original 30 seconds). The closed preview body permits 64 MiB under the original 30-second clock. Whole or selected physical roots yield complete freshly qualified versioned dependencies and pins, plus all affected derivations. Page sections are included, dependencies, relations or conflicts; offset 0..50000 and limit 1..100. All pages share the complete preview digest and pins. Apply explicit conflict, value and placement choices through the existing compose-profile command; donor values are never imported.",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "parameters": [
          {
            "name": "planId",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Uuid"
            }
          },
          {
            "name": "X-CSRF-TOKEN",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "../../schemas/plan-workflow-v3.schema.json#/$defs/previewRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Owned complete result or explicit computed-rule page; no export authority",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "../../schemas/plan-workflow-v3.schema.json#/$defs/previewResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error400"
          },
          "401": {
            "$ref": "#/components/responses/Error401"
          },
          "403": {
            "$ref": "#/components/responses/Error403"
          },
          "404": {
            "$ref": "#/components/responses/Error404"
          },
          "409": {
            "$ref": "#/components/responses/Error409"
          },
          "413": {
            "$ref": "#/components/responses/Error413"
          },
          "422": {
            "$ref": "#/components/responses/Error422"
          },
          "429": {
            "$ref": "#/components/responses/Error429"
          },
          "503": {
            "$ref": "#/components/responses/Error503"
          },
          "500": {
            "$ref": "#/components/responses/Error500"
          }
        }
      }
    },
    "/api/v3/plans/{planId}/validations": {
      "post": {
        "operationId": "validatePlanV3",
        "summary": "Evaluate complete v3 validation and inspect fingerprint-bound rule pages",
        "description": "See hosted-plan-workflow-v3.md. Original V3 owner, full view admission and the shared semantic transfer precede input; original authority and full proofs remain checked through owned output (128 MiB, original 30 seconds). The closed validation body permits 16 KiB under the original 10-second clock. A revision-only request returns complete checks, physical rules, target completeness and nullable computed-rule count without the full computed-rule array. Computed-rule page requests require the exact validation inputFingerprint, offset 0..2147483647 and limit 1..100. Fingerprint mismatch refuses CONFLICT; missing target refuses INCOMPLETE_TARGET. An oversized page refuses RESOURCE_LIMIT; request fewer rows at the same offset and fingerprint. Counts remain complete and cardinalities are exact decimal strings. Export remains unavailable.",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "parameters": [
          {
            "name": "planId",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Uuid"
            }
          },
          {
            "name": "X-CSRF-TOKEN",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "../../schemas/plan-workflow-v3.schema.json#/$defs/validationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Owned complete result or explicit computed-rule page; no export authority",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "../../schemas/plan-workflow-v3.schema.json#/$defs/validationResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error400"
          },
          "401": {
            "$ref": "#/components/responses/Error401"
          },
          "403": {
            "$ref": "#/components/responses/Error403"
          },
          "404": {
            "$ref": "#/components/responses/Error404"
          },
          "409": {
            "$ref": "#/components/responses/Error409"
          },
          "413": {
            "$ref": "#/components/responses/Error413"
          },
          "422": {
            "$ref": "#/components/responses/Error422"
          },
          "429": {
            "$ref": "#/components/responses/Error429"
          },
          "503": {
            "$ref": "#/components/responses/Error503"
          },
          "500": {
            "$ref": "#/components/responses/Error500"
          }
        }
      }
    },
...
    "/api/v3/plans/{planId}/reviews": {
      "post": {
        "operationId": "acknowledgePlanReviewV3",
        "summary": "Acknowledge exact checked v3 plan inputs without granting export",
        "description": "See plan-review-v3.md. Original V3 lease/view and shared semantic transfer precede input. Closed 16 KiB body under original 10-second deadline; 32 KiB acknowledgement under original 30-second output deadline. Fresh publication and complete XML proof bind exact fingerprint, destination and protected-self-contained intent. Shared 256-entry replay is checked before current revision pinning and never reactivates review. Original async abort is atomically ordered before commit; response loss after commit preserves the historical receipt. Review and immutable whole-document policy checks do not qualify client capability or authorize export.",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "parameters": [
          {
            "name": "planId",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Uuid"
            }
          },
          {
            "name": "X-CSRF-TOKEN",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "../../schemas/plan-review-v3.schema.json#/$defs/request"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Original unchanged-revision receipt; no operationId or export authority",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "../../schemas/plan-review-v3.schema.json#/$defs/response"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error400"
          },
          "401": {
            "$ref": "#/components/responses/Error401"
          },
          "403": {
            "$ref": "#/components/responses/Error403"
          },
          "404": {
            "$ref": "#/components/responses/Error404"
          },
          "409": {
            "$ref": "#/components/responses/Error409"
          },
          "413": {
            "$ref": "#/components/responses/Error413"
          },
          "422": {
            "$ref": "#/components/responses/Error422"
          },
          "429": {
            "$ref": "#/components/responses/Error429"
          },
          "503": {
            "$ref": "#/components/responses/Error503"
          },
          "500": {
            "$ref": "#/components/responses/Error500"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Uuid": {
        "type": "string",
        "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}(?![\\s\\S])"
      },
...
      "Code": {
        "type": "string",
        "pattern": "^[A-Z][A-Z0-9_]{0,63}(?![\\s\\S])"
      },
...
      "Error": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "code"
        ],
        "properties": {
          "code": {
            "$ref": "#/components/schemas/Code"
          }
        }
      },
...
      "ControllerErrorCode": {
        "type": "string",
        "enum": [
          "BODY_DEADLINE",
          "BODY_TOO_LARGE",
          "CANCELLED",
          "CAPACITY",
          "CLEANUP_INCONCLUSIVE",
          "CONFLICT",
          "CREDENTIALS_ALREADY_CONSUMED",
          "DESTINATION_DENIED",
          "DISCLOSURE_REQUIRED",
          "EXPORT_UNAVAILABLE",
          "INCOMPLETE_TARGET",
          "INSPECTION_REQUIRED",
          "INVALID_CREDENTIALS",
          "INVALID_DESTINATION",
          "INVALID_REQUEST",
          "MALFORMED_BODY",
          "NOT_FOUND",
          "OBSERVATION_REFUSED",
          "PLAN_BUSY",
          "PLAN_INTERNAL_REFUSAL",
          "PLAN_SERVICES_UNAVAILABLE",
          "PROFILE_REFUSED",
          "PROJECTION_REFUSED",
          "PUBLICATION_REQUIRED",
          "RESERVATION_EXPIRED",
          "RESOURCE_LIMIT",
          "SESSION_REQUIRED",
          "STALE_PREVIEW",
          "UNSUPPORTED_DEFINITION"
        ]
      },
...
    },
    "responses": {
      "Error400": {
        "description": "Malformed or unknown request fields; an early controller refusal has no body and includes Content-Length: 0 plus X-Environment-Studio-Code together. The closed-code header denotes that branch and requires zero length. Other branches may also carry a nonzero Content-Length. An owned transfer or security middleware may instead return its documented JSON error body.",
        "headers": {
          "Cache-Control": {
            "schema": {
              "type": "string",
              "const": "no-store"
            }
          },
          "Content-Length": {
            "description": "Zero when X-Environment-Studio-Code denotes an early no-body controller refusal. Owned JSON errors carry their actual encoded length. OpenAPI cannot enforce this conditional relationship between headers and body.",
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          },
          "X-Environment-Studio-Code": {
            "description": "Closed safe code on the early no-body controller branch; paired with Content-Length: 0.",
            "schema": {
              "$ref": "#/components/schemas/ControllerErrorCode"
            }
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Error401": {
        "description": "No live hosted authentication; an early controller refusal has no body and includes Content-Length: 0 plus X-Environment-Studio-Code together. The closed-code header denotes that branch and requires zero length. Other branches may also carry a nonzero Content-Length. An owned transfer or security middleware may instead return its documented JSON error body.",
        "headers": {
          "Cache-Control": {
            "schema": {
              "type": "string",
              "const": "no-store"
            }
          },
          "Content-Length": {
            "description": "Zero when X-Environment-Studio-Code denotes an early no-body controller refusal. Owned JSON errors carry their actual encoded length. OpenAPI cannot enforce this conditional relationship between headers and body.",
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          },
          "X-Environment-Studio-Code": {
            "description": "Closed safe code on the early no-body controller branch; paired with Content-Length: 0.",
            "schema": {
              "$ref": "#/components/schemas/ControllerErrorCode"
            }
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Error403": {
        "description": "Host, Origin, CSRF or policy denied; unavailable in demo; an early controller refusal has no body and includes Content-Length: 0 plus X-Environment-Studio-Code together. The closed-code header denotes that branch and requires zero length. Other branches may also carry a nonzero Content-Length. An owned transfer or security middleware may instead return its documented JSON error body.",
        "headers": {
          "Cache-Control": {
            "schema": {
              "type": "string",
              "const": "no-store"
            }
          },
          "Content-Length": {
            "description": "Zero when X-Environment-Studio-Code denotes an early no-body controller refusal. Owned JSON errors carry their actual encoded length. OpenAPI cannot enforce this conditional relationship between headers and body.",
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          },
          "X-Environment-Studio-Code": {
            "description": "Closed safe code on the early no-body controller branch; paired with Content-Length: 0.",
            "schema": {
              "$ref": "#/components/schemas/ControllerErrorCode"
            }
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Error404": {
        "description": "Missing, foreign-owner, foreign-lease or wrong-version resource; original retained v3 operation metadata remains eligible; an early controller refusal has no body and includes Content-Length: 0 plus X-Environment-Studio-Code together. The closed-code header denotes that branch and requires zero length. Other branches may also carry a nonzero Content-Length. An owned transfer or security middleware may instead return its documented JSON error body.",
        "headers": {
          "Cache-Control": {
            "schema": {
              "type": "string",
              "const": "no-store"
            }
          },
          "Content-Length": {
            "description": "Zero when X-Environment-Studio-Code denotes an early no-body controller refusal. Owned JSON errors carry their actual encoded length. OpenAPI cannot enforce this conditional relationship between headers and body.",
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          },
          "X-Environment-Studio-Code": {
            "description": "Closed safe code on the early no-body controller branch; paired with Content-Length: 0.",
            "schema": {
              "$ref": "#/components/schemas/ControllerErrorCode"
            }
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Error409": {
        "description": "Stale or colliding request, busy plan or consumed reservation; an early controller refusal has no body and includes Content-Length: 0 plus X-Environment-Studio-Code together. The closed-code header denotes that branch and requires zero length. Other branches may also carry a nonzero Content-Length. An owned transfer or security middleware may instead return its documented JSON error body.",
        "headers": {
          "Cache-Control": {
            "schema": {
              "type": "string",
              "const": "no-store"
            }
          },
          "Content-Length": {
            "description": "Zero when X-Environment-Studio-Code denotes an early no-body controller refusal. Owned JSON errors carry their actual encoded length. OpenAPI cannot enforce this conditional relationship between headers and body.",
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          },
          "X-Environment-Studio-Code": {
            "description": "Closed safe code on the early no-body controller branch; paired with Content-Length: 0.",
            "schema": {
              "$ref": "#/components/schemas/ControllerErrorCode"
            }
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Error413": {
        "description": "Wire byte limit exceeded; an early controller refusal has no body and includes Content-Length: 0 plus X-Environment-Studio-Code together. The closed-code header denotes that branch and requires zero length. Other branches may also carry a nonzero Content-Length. An owned transfer or security middleware may instead return its documented JSON error body.",
        "headers": {
          "Cache-Control": {
            "schema": {
              "type": "string",
              "const": "no-store"
            }
          },
          "Content-Length": {
            "description": "Zero when X-Environment-Studio-Code denotes an early no-body controller refusal. Owned JSON errors carry their actual encoded length. OpenAPI cannot enforce this conditional relationship between headers and body.",
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          },
          "X-Environment-Studio-Code": {
            "description": "Closed safe code on the early no-body controller branch; paired with Content-Length: 0.",
            "schema": {
              "$ref": "#/components/schemas/ControllerErrorCode"
            }
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Error422": {
        "description": "Semantic operation refused; an early controller refusal has no body and includes Content-Length: 0 plus X-Environment-Studio-Code together. The closed-code header denotes that branch and requires zero length. Other branches may also carry a nonzero Content-Length. An owned transfer or security middleware may instead return its documented JSON error body.",
        "headers": {
          "Cache-Control": {
            "schema": {
              "type": "string",
              "const": "no-store"
            }
          },
          "Content-Length": {
            "description": "Zero when X-Environment-Studio-Code denotes an early no-body controller refusal. Owned JSON errors carry their actual encoded length. OpenAPI cannot enforce this conditional relationship between headers and body.",
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          },
          "X-Environment-Studio-Code": {
            "description": "Closed safe code on the early no-body controller branch; paired with Content-Length: 0.",
            "schema": {
              "$ref": "#/components/schemas/ControllerErrorCode"
            }
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Error429": {
        "description": "Bounded plan, operation or memory capacity exhausted; an early controller refusal has no body and includes Content-Length: 0 plus X-Environment-Studio-Code together. The closed-code header denotes that branch and requires zero length. Other branches may also carry a nonzero Content-Length. An owned transfer or security middleware may instead return its documented JSON error body.",
        "headers": {
          "Cache-Control": {
            "schema": {
              "type": "string",
              "const": "no-store"
            }
          },
          "Content-Length": {
            "description": "Zero when X-Environment-Studio-Code denotes an early no-body controller refusal. Owned JSON errors carry their actual encoded length. OpenAPI cannot enforce this conditional relationship between headers and body.",
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          },
          "X-Environment-Studio-Code": {
            "description": "Closed safe code on the early no-body controller branch; paired with Content-Length: 0.",
            "schema": {
              "$ref": "#/components/schemas/ControllerErrorCode"
            }
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Error503": {
        "description": "Trusted service or configuration unavailable; an early controller refusal has no body and includes Content-Length: 0 plus X-Environment-Studio-Code together. The closed-code header denotes that branch and requires zero length. Other branches may also carry a nonzero Content-Length. An owned transfer or security middleware may instead return its documented JSON error body.",
        "headers": {
          "Cache-Control": {
            "schema": {
              "type": "string",
              "const": "no-store"
            }
          },
          "Content-Length": {
            "description": "Zero when X-Environment-Studio-Code denotes an early no-body controller refusal. Owned JSON errors carry their actual encoded length. OpenAPI cannot enforce this conditional relationship between headers and body.",
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          },
          "X-Environment-Studio-Code": {
            "description": "Closed safe code on the early no-body controller branch; paired with Content-Length: 0.",
            "schema": {
              "$ref": "#/components/schemas/ControllerErrorCode"
            }
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Error500": {
        "description": "Closed internal controller refusal; an early controller refusal has no body and includes Content-Length: 0 plus X-Environment-Studio-Code together. The closed-code header denotes that branch and requires zero length. Other branches may also carry a nonzero Content-Length. An owned transfer or security middleware may instead return its documented JSON error body.",
        "headers": {
          "Cache-Control": {
            "schema": {
              "type": "string",
              "const": "no-store"
            }
          },
          "Content-Length": {
            "description": "Zero when X-Environment-Studio-Code denotes an early no-body controller refusal. Owned JSON errors carry their actual encoded length. OpenAPI cannot enforce this conditional relationship between headers and body.",
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          },
          "X-Environment-Studio-Code": {
            "description": "Closed safe code on the early no-body controller branch; paired with Content-Length: 0.",
            "schema": {
              "$ref": "#/components/schemas/ControllerErrorCode"
            }
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    },
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "JSESSIONID"
      }
    }
  }
}